move `toBigNumber` to `things.functions.js`
don't repeat yourself don't repeat yourself don't repeat yourself don't repeat yourself don't repeat yourself don't repeat yourself don't repeat yourself don't repeat yourself don't repeat yourself don't repeat yourself
