+++ listeners/exp.js
@@ -3,87 +3,7 @@
 const {Listener} = require("discord-akairo");
 const UserModel = require("../model.user.js");
 
-const {expNeededForLevel} = require.main.require("./things.functions.js");
-
-function toBigNumber(number) {
-	/* array number corresponds with number. (bigNumbers[0] is zero)
-       object id corresponds with line number (bigNumbers[0].2 is line 2 of zero.). */
-	const bigNumbers = [[
-		"██████ ",
-		"██  ██ ",
-		"██  ██ ",
-		"██  ██ ",
-		"██████ "
-	], [
-		"████   ",
-		"  ██   ",
-		"  ██   ",
-		"  ██   ",
-		"██████ "
-	], [
-		"██████ ",
-		"    ██ ",
-		"██████ ",
-		"██     ",
-		"██████ "
-	], [
-		"██████ ",
-		"    ██ ",
-		"██████ ",
-		"    ██ ",
-		"██████ "
-	], [
-		"██  ██ ",
-		"██  ██ ",
-		"██████ ",
-		"    ██ ",
-		"    ██ "
-	], [
-		"██████ ",
-		"██     ",
-		"██████ ",
-		"    ██ ",
-		"██████ "
-	], [
-		"██████ ",
-		"██     ",
-		"██████ ",
-		"██  ██ ",
-		"██████ "
-	], [
-		"██████ ",
-		"    ██ ",
-		"    ██ ",
-		"    ██ ",
-		"    ██ "
-	], [
-		"██████ ",
-		"██  ██ ",
-		"██████ ",
-		"██  ██ ",
-		"██████ "
-	], [
-		"██████ ",
-		"██  ██ ",
-		"██████ ",
-		"    ██ ",
-		"██████ "
-	]];
-	// turn number into array of numbers
-	const numberDigits = Array.from(String(number)).map(Number);
-
-	let output = "";
-
-	for (let line = 0; line < 5; line++) {
-		for (let digit = 0; digit < numberDigits.length; digit++) {
-			output += bigNumbers[numberDigits[digit]][line];
-		}
-
-		output += "\n";
-	}
-
-	return output;
-}
+const {expNeededForLevel, toBigNumber} = require.main.require("./things.functions.js");
 
 class ExperienceListener extends Listener {
 	constructor() {
